+++ common/src/main/kotlin/com/catchrate/CatchRateCalculator.kt
@@ -150,7 +150,7 @@ object CatchRateCalculator {
     }
 
     private fun buildFallbackResult(pokemon: ClientBattlePokemon, ballName: String, turnCount: Int): CatchRateResult {
-        val baseCatchRate = try { SpeciesCatchRateCache.getCatchRate(pokemon.species) } catch (_: Throwable) { 45 }
+        val baseCatchRate = try { SpeciesCatchRateCache.getCatchRate(pokemon.species) } catch (_: Throwable) { SpeciesCatchRateCache.fallbackCatchRate() }
         return CatchRateResult(
             percentage = CatchRateFormula.modifiedRateToPercentage(baseCatchRate.toFloat()).toDouble().coerceIn(0.0, 100.0),
             hpPercentage = 100.0,